requirements.txt
Django==3.0.7
djangorestframework==3.11.0
Pillow==7.1.2
django-rest-registration==0.5.6